+++ src/main/java/net/collaud/fablab/manager/dao/AssetRepository.java
@@ -8,7 +8,7 @@ import org.springframework.transaction.annotation.Transactional;
  * @author Gaetan Collaud <gaetancollaud@gmail.com>
  */
 @Transactional
-public interface AssetRepository extends JpaRepository<AssetEO, Integer>, AssetRepositoryCustom{
+public interface AssetRepository extends JpaRepository<AssetEO, Long>, AssetRepositoryCustom{
