+++ src/main/java/net/collaud/fablab/manager/dao/AuditRepository.java
@@ -9,7 +9,7 @@ import org.springframework.transaction.annotation.Transactional;
  * @author Gaetan Collaud <gaetancollaud@gmail.com>
  */
 @Transactional
-public interface AuditRepository extends JpaRepository<AuditEO, Integer>, AuditRepositoryCustom{
+public interface AuditRepository extends JpaRepository<AuditEO, Long>, AuditRepositoryCustom{
 
 
 }
