+++ src/main/java/net/collaud/fablab/manager/dao/AuditRepositoryCustom.java
@@ -12,6 +12,6 @@ import net.collaud.fablab.manager.exceptions.FablabException;
  */
 public interface AuditRepositoryCustom{
 
-	public List<AuditEO> search(Integer userId, List<AuditObject> type, Date after, Date before, String content, int limit) throws FablabException;
+	public List<AuditEO> search(Long userId, List<AuditObject> type, Date after, Date before, String content, int limit) throws FablabException;
 
 }
