+++ src/main/java/net/collaud/fablab/manager/data/PriceMachineEO.java
@@ -4,7 +4,6 @@ import com.fasterxml.jackson.annotation.JsonBackReference;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import java.io.Serializable;
 import javax.persistence.Column;
-import javax.persistence.EmbeddedId;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.Id;
@@ -12,7 +11,6 @@ import javax.persistence.IdClass;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
-import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
@@ -31,14 +29,14 @@ public class PriceMachineEO extends AbstractDataEO<PriceMachineEOPK> implements 
 
 	@Id
 	@Column(name = "machine_type_id", nullable = false)
-	private int machineTypeId;
+	private Long machineTypeId;
 
 	@Id
 	@Column(name = "membership_type_id", nullable = false)
-	private int membershipTypeId;
+	private Long membershipTypeId;
 
 	@Column(name = "price", nullable = false)
-	private float price;
+	private Float price;
 
 	@JsonBackReference("machineType-price")
 	@JoinColumn(name = "machine_type_id", referencedColumnName = "machine_type_id", insertable = false, updatable = false)
