+++ src/main/java/net/collaud/fablab/manager/service/SecurityService.java
@@ -18,7 +18,7 @@ public interface SecurityService {
 	
 	Optional<UserEO> getCurrentUser();
 	
-	Integer getCurrentUserId();
+	Long getCurrentUserId();
 	
 	Boolean isAuthenticated();
